requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
groq>=0.4.0
duckduckgo-search>=6.0.0
googlesearch-python>=1.2.0
python-dotenv>=1.0.0
